refactor: extract the wizard backdrop and state mappers from the creation wizard
Move the 5-layer decorative backdrop into WizardBackdrop, and relocate
buildDefaults/autopilotToWizardState into features/autopilot/lib/wizard-state so
AutopilotPage imports the mapper from there rather than the component. index.tsx
keeps the wizard shell.
