+++ apps/tauri/src/renderer/features/autopilot/components/AutopilotPage/index.tsx
@@ -6,12 +6,10 @@ import { Button } from '@ajh/ui';
 
 import { PageTransition } from '@/components/layout/PageTransition';
 import { AutopilotCard } from '@/features/autopilot/components/AutopilotCard';
-import {
-  autopilotToWizardState,
-  CreationWizard,
-} from '@/features/autopilot/components/CreationWizard';
+import { CreationWizard } from '@/features/autopilot/components/CreationWizard';
 import { EmptyState } from '@/features/autopilot/components/EmptyState';
 import { useAutopilotRun } from '@/features/autopilot/hooks/useAutopilotRun';
+import { autopilotToWizardState } from '@/features/autopilot/lib/wizard-state';
 import { useTranslation } from '@/lib/i18n';
 import { useAutopilots } from '@/services';
 import { useSessionStore } from '@/store/session-store';
