+++ apps/tauri/src/renderer/features/autopilot/components/CreationWizard/index.tsx
@@ -9,10 +9,13 @@ import { StepAction } from '@/features/autopilot/components/wizard-steps/StepAct
 import { StepFilter } from '@/features/autopilot/components/wizard-steps/StepFilter';
 import { StepSchedule } from '@/features/autopilot/components/wizard-steps/StepSchedule';
 import { StepTarget } from '@/features/autopilot/components/wizard-steps/StepTarget';
+import { buildDefaults } from '@/features/autopilot/lib/wizard-state';
 import type { SetFn, WizardState } from '@/features/autopilot/types';
 import { useTranslation } from '@/lib/i18n';
 import { useCreateAutopilot, useJobPreferences, useUpdateAutopilot } from '@/services';
 import { useSessionStore } from '@/store/session-store';
+
+import { WizardBackdrop } from './WizardBackdrop';
 
 interface CreationWizardProps {
   onDone(ap: Autopilot): void;
@@ -20,50 +23,6 @@ interface CreationWizardProps {
 }
 
 const STEPS = ['Target', 'Filter', 'Action', 'Schedule'] as const;
-
-function buildDefaults(jobPrefs?: JobPreferences): WizardState {
-  const validWorkType = ['remote', 'hybrid', 'on-site', 'any'] as const;
-  return {
-    name: '',
-    board: 'linkedin',
-    query: '',
-    location: jobPrefs?.location ?? '',
-    workType: validWorkType.includes(jobPrefs?.remote as (typeof validWorkType)[number])
-      ? (jobPrefs?.remote as WizardState['workType'])
-      : 'any',
-    pages: 2,
-    dateFilter: '24h',
-    minMatchScore: 50,
-    keywords: jobPrefs?.techStack?.map((t) => t.name).join(', ') ?? '',
-    excludeKeywords: '',
-    resumeText: '',
-    action: 'save',
-    coverLetter: '',
-    autoSubmit: false,
-    schedule: 'daily',
-  };
-}
-
-/** Map a persisted autopilot back into the wizard form for editing. */
-export function autopilotToWizardState(ap: Autopilot): WizardState {
-  return {
-    name: ap.name,
-    board: ap.target.board,
-    query: ap.target.query,
-    location: ap.target.location ?? '',
-    workType: ap.target.workType ?? 'any',
-    pages: ap.target.pages,
-    dateFilter: ap.target.dateFilter ?? '',
-    minMatchScore: ap.filter.minMatchScore,
-    keywords: ap.filter.keywords?.join(', ') ?? '',
-    excludeKeywords: ap.filter.excludeKeywords?.join(', ') ?? '',
-    resumeText: ap.resumeText ?? '',
-    action: ap.action,
-    coverLetter: ap.coverLetter ?? '',
-    autoSubmit: ap.autoSubmit,
-    schedule: ap.schedule,
-  };
-}
 
 export function CreationWizard({ onDone, onCancel }: CreationWizardProps) {
   const { t } = useTranslation();
@@ -163,89 +122,9 @@ export function CreationWizard({ onDone, onCancel }: CreationWizardProps) {
       transition={transition.fast}
       className="absolute inset-0 z-[var(--z-modal)] flex items-center justify-center"
     >
-      {/*
-       * BACKDROP LAYERING — reading order = bottom to top
-       *
-       * 1. blur layer     — destroys shape detail
-       * 2. crush layer    — kills remaining contrast/readability
-       * 3. bokeh blobs    — adds ambient atmospheric color
-       * 4. vignette       — pulls focus to center
-       * 5. modal panel    — the actual foreground
-       */}
+      <WizardBackdrop />
 
-      {/* 1 — heavy blur. backdrop-filter alone preserves contrast,
-              so we follow it with a dark crush layer. */}
-      <div
-        className="absolute inset-0"
-        style={{
-          backdropFilter: 'blur(64px) saturate(120%) brightness(0.6)',
-          WebkitBackdropFilter: 'blur(64px) saturate(120%) brightness(0.6)',
-        }}
-      />
-
-      {/* 2 — contrast crush: semi-opaque dark tint that makes the
-              blurred background unreadable without going full black. */}
-      <div className="absolute inset-0 bg-background/70" />
-
-      {/* 3 — bokeh blobs: soft ambient color on top of the dark crush. */}
-      <div className="absolute inset-0 pointer-events-none overflow-hidden">
-        <motion.div
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={transition.blob1}
-          className="absolute"
-          style={{
-            top: '-5%',
-            left: '5%',
-            width: 500,
-            height: 500,
-            background: 'radial-gradient(circle, rgba(168,85,247,0.18) 0%, transparent 65%)',
-            filter: 'blur(48px)',
-          }}
-        />
-        <motion.div
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={transition.blob2}
-          className="absolute"
-          style={{
-            bottom: '0%',
-            right: '5%',
-            width: 420,
-            height: 420,
-            background: 'radial-gradient(circle, rgba(79,70,229,0.15) 0%, transparent 65%)',
-            filter: 'blur(56px)',
-          }}
-        />
-        <motion.div
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={transition.blob3}
-          className="absolute"
-          style={{
-            top: '40%',
-            left: '40%',
-            width: 300,
-            height: 300,
-            background: 'radial-gradient(circle, rgba(192,38,211,0.07) 0%, transparent 65%)',
-            filter: 'blur(72px)',
-          }}
-        />
-      </div>
-
-      {/* 4 — vignette: darkens screen edges, naturally centres the eye. */}
-      <div
-        className="absolute inset-0 pointer-events-none"
-        style={{
-          background:
-            'radial-gradient(ellipse 80% 80% at 50% 50%, transparent 35%, rgba(0,0,0,0.65) 100%)',
-        }}
-      />
-
-      {/* 5 — modal glass panel */}
+      {/* Modal glass panel */}
       <motion.div
         initial={{ opacity: 0, scale: 0.96, y: 14 }}
         animate={{ opacity: 1, scale: 1, y: 0 }}
